requirements.txt
#nvidia-cudnn-cu11 
cellpose
#torch==2.0.1
#torchvision==0.15.2
cupy-cuda11x
cucim-cu11
dask[complete]
scikit-image
scikit-learn
imagecodecs
opencv-python-headless
scipy
numpy
matplotlib
seaborn
trimesh
multiprocess
tiler
connected-components-3d
edt
tqdm
scikit-fmm
pyicu
czifile
